Rewrite the main import check test.
The purpose of the rewrite is to test the full data importer by the
password store importer. Only the password and login fields used to be
checked.

This new test structure is required because a lot of pm do not have
the field tested. Therefore, the test system need to update the
reference data accordingly.