Handle entries with empty tags in passman manager
Attempting to access the first element of an empty list will
`throw IndexError`. Tags are not a mandatory field for password entries
in PassMan and hence there can be entries without tags as well.

This situation would previously crash with
`IndexError: list index out of range`
obviously because of the attempt to access the zeroth element of an
empty list. This change fixes that situation.