Raise an exception upon invalid LastPass CSV data
I have encountered various scenarios where LastPass will truncate
entries in the CSV file it generates during an export, leading to
a CSV file with invalid rows!

In many of these cases, the result is that the "group" field will
appear to be empty, which is not correct and hence was not expected by
LastpassCSV. Here we don't attempt to ignore the issue, but instead
provide a slightly more helpful error message when we fail.